feat(nombre_decimal): On peut saisir des montant décimal, ça fonctionne

+++ app/src/main/java/com/fsp/ticket_resto_app/controlleur/Controlleur.java
@@ -32,7 +32,7 @@ public class Controlleur {
         return instance;
     }
 
-    public void setMontantAPayer(int montant) {
+    public void setMontantAPayer(double montant) {
         decomposition.setMontant(montant);
     }
